# Bug 24185: Make holds page fast when 'on shelf holds' set to 'If all unavailable'